backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
openai>=1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.2
python-dotenv==1.0.1
Appium-Python-Client>=3.0.0
apkutils2>=1.0.0
schedule>=1.2.0
